requirements.txt
alabaster==0.7.12
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bleach==3.1.5
bottle==0.12.18
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.4.1
colorama==0.4.3
construct==2.10.56
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
docutils==0.16
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
helpdev==0.7.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
matplotlib-venn==0.11.5
mccabe==0.6.1
mistune==0.8.4
mod-wsgi==4.7.1+ap24vc14
nbconvert==5.6.1
nbformat==5.0.6
numpy==1.18.4
numpydoc==0.9.2
packaging==20.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.7.2
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==19.0.0
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
rope==0.17.0
scipy==1.4.1
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.2
spyder-kernels==1.9.1
sqlparse==0.4.1
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yapf==0.30.0
zipp==3.1.0
